ATT-4: move Upload to AttachmentResource1_10
